YCR: decompose server root mutation
Basically moves all the implementation details under `auth` where it belongs.
